Add files via upload
The new csv file is needed for "part two" of the "Run File". Part one relies on the already uploaded "....MK_bearbeitet2.csv" and part three creates its own trace. Of course this is only relevant if you want to try out the current state, in the future "all parts" should work together with one csv file and regardless which one you use.